# docs/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SecureHealth Enterprise Platform API
  description: |
    Enterprise-grade healthcare platform with microservices architecture.
    
    ## Architecture Overview
    - **Patient Service**: User profiles, wallet metadata, consent records
    - **Hospital Service**: Hospital admin, document signing, EHR adapters  
    - **Document Service**: File ingestion, encryption, storage, hash calculation
    - **Blockchain Adapter**: Signs/writes hashes, manages anchor transactions
    - **Claims Service**: Insurer integrations, adjudication workflows
    - **Auth Service**: ABHA/Aadhaar integrations, OAuth2, RBAC
    
    ## Key Design Principles
    - Never store PHI on-chain (only cryptographic hashes)
    - Multi-tenant isolation per hospital/insurer
    - Event-driven architecture with Kafka
    - Hybrid blockchain (permissioned + public anchor)
  version: 1.0.0
  contact:
    name: SecureHealth API Team
    email: api@securehealth.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.securehealth.app/v1
    description: Production API Gateway
  - url: https://staging-api.securehealth.app/v1
    description: Staging Environment
  - url: http://localhost:3000/v1
    description: Local Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /hospitals:
    get:
      tags: [Hospitals]
      summary: List hospitals (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of hospitals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HospitalListResponse'

    post:
      tags: [Hospitals]
      summary: Register new hospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HospitalRegistrationRequest'
      responses:
        '201':
          description: Hospital registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HospitalResponse'
  /hospitals/{hospitalId}:
    get:
      tags: [Hospitals]
      summary: Get hospital details
      parameters:
        - name: hospitalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hospital details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HospitalResponse'
  /hospitals/{hospitalId}/patients/search:
    post:
      tags: [Hospitals]
      summary: Search patients by ABHA/phone/name
      parameters:
        - name: hospitalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientSearchRequest'
      responses:
        '200':
          description: Patient search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientSearchResponse'
  /hospitals/{hospitalId}/documents:
    post:
      tags: [Hospitals]
      summary: Upload and sign document
      parameters:
        - name: hospitalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (PDF, image, etc.)
                patientId:
                  type: string
                  description: ABHA ID or patient identifier
                documentType:
                  type: string
                  enum: [discharge_summary, lab_report, prescription, imaging, pathology, other]
                metadata:
                  type: string
                  description: JSON metadata for the document
              required: [file, patientId, documentType]
      responses:
        '201':
          description: Document uploaded and signed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'

  # Document Service Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    HospitalRegistrationRequest:
      type: object
      properties:
        name:
          type: string
        registrationNumber:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        contactInfo:
          $ref: '#/components/schemas/ContactInfo'
        specialties:
          type: array
          items:
            type: string
        kycDocuments:
          type: array
          items:
            $ref: '#/components/schemas/KYCDocument'
      required: [name, registrationNumber, address, contactInfo]
    HospitalResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        registrationNumber:
          type: string
        status:
          type: string
          enum: [pending, verified, suspended]
        address:
          $ref: '#/components/schemas/Address'
        contactInfo:
          $ref: '#/components/schemas/ContactInfo'
        specialties:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
    HospitalListResponse:
      type: object
      properties:
        hospitals:
          type: array
          items:
            $ref: '#/components/schemas/HospitalResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Patient Schemas
    PatientSearchRequest:
      type: object
      properties:
        abhaId:
          type: string
        phone:
          type: string
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
      oneOf:
        - required: [abhaId]
        - required: [phone]
        - required: [name, dateOfBirth]
    PatientSearchResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: '#/components/schemas/PatientProfile'
        total:
          type: integer
    PatientProfile:
      type: object
      properties:
        id:
          type: string
        abhaId:
          type: string
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        phone:
          type: string
        email:
          type: string
          format: email
        address:
          $ref: '#/components/schemas/Address'
        emergencyContact:
          $ref: '#/components/schemas/EmergencyContact'
...
    DocumentUploadResponse:
      type: object
      properties:
        documentId:
          type: string
        fileHash:
          type: string
        ipfsHash:
          type: string
        blockchainTx:
          type: string
        url:
          type: string
        metadata:
          $ref: '#/components/schemas/DocumentMetadata'
    DocumentMetadata:
      type: object
      properties:
        documentId:
          type: string
        patientId:
          type: string
        hospitalId:
          type: string
        type:
          type: string
          enum: [discharge_summary, lab_report, prescription, imaging, pathology, other]
        fileHash:
          type: string
        ipfsHash:
          type: string
        blockchainTx:
          type: string
        signedBy:
          type: string
        signature:
          type: string
        uploadedAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, verified, rejected, expired]
        size:
          type: integer
        mimeType:
          type: string
        originalName:
          type: string

    # Consent Schemas
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
          default: India
    ContactInfo:
      type: object
      properties:
        phone:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
    EmergencyContact:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
    KYCDocument:
      type: object
      properties:
        type:
          type: string
          enum: [registration_certificate, license, pan_card, gst_certificate]
        url:
          type: string
        verified:
          type: boolean
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...